requirements.txt
# --- Core System & Data ---
pandas==2.3.0
numpy==2.3.5
joblib==1.5.1
PyYAML==6.0.3
requests==2.32.4
python-dotenv==1.2.1# --- Core System & Data ---
pandas==2.2.3
numpy==1.26.4
joblib==1.3.2
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.1
pathlib

# --- Machine Learning Models ---
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1

# --- Validation & Quality ---
pandera==0.17.2
pydantic==2.6.4

# --- Serving & API ---
fastapi==0.110.0
uvicorn==0.29.0
prometheus-fastapi-instrumentator==6.1.0

# --- MLOps & Tracking ---
mlflow==2.12.1
prefect==2.14.20
evidently==0.3.3

# --- Explainability ---
shap==0.42.1
matplotlib==3.8.3

# --- Database (Postgres) ---
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

pathlib

# --- Machine Learning Models ---
scikit-learn==1.7.0
xgboost==3.1.2
lightgbm==4.6.0
optuna==4.7.0

# --- Validation & Quality ---
pandera==0.29.0
pydantic==2.12.5

# --- Serving & API ---
fastapi==0.116.0
uvicorn==0.35.0
prometheus-fastapi-instrumentator==7.1.0

# --- MLOps & Tracking ---
mlflow==3.11.1
prefect==3.6.11
evidently==0.4.15


# --- Explainability ---
shap==0.50.0
matplotlib==3.10.8

# --- Database (Postgres) ---
psycopg2-binary==2.9.11
SQLAlchemy==2.0.45